requirements.txt
python-telegram-bot==20.8
gspread
oauth2client
httpx==0.27.0
google-api-python-client==2.137.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
yookassa==3.0.0
tzdata>=2024.1